new hash map implementation: compact open addressing
this commit introduces a new implementation of open addressing with a more compact representation. It does this by separating out the status bits and the actual data so that
the data fits inside a (u64,u64) tuple and takes up 16 bytes total. This should fit better in a cache line of 64 bytes (4 per cache line) versus the earlier structure which
was 24 bytes (2 per cache line at most)